requirements.txt
Flask==3.0.0
stripe==7.10.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
werkzeug==3.0.1
PyJWT==2.8.0
supabase==2.3.0
